Move libvirt connection mixin
